policy: do not rebuild policy package if already built
Right now, test_policy.pp is rebuilt on every make invocation. Tweak the
Makefile so that the concatenated .te file is only regenerated when it
doesn't exist. The POLDEV Makefile will then take care of rebuilding the
package as needed (when test_policy.te or POLDEV files are updated).

This patch also disables building policy in Travis - it was failing
silently before (no POLDEV files in the CI environment), now it's
actually failing the build.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>