tests/sctp: really fix the ASCONF test race condition
Commit 8eac676ec2dc ("tests/sctp: fix a race condition in the new ASCONF
test") attempted to fix the race condition by adding a sleep() call, but
this wasn't enough. Running the testsuite on a realtime kernel in a VM
uncovered that the race can still happen, leading to frequent failures.

To fix it properly, unroll the client's receiving while loop and add a
synchronization point between client and server by sending a message
client->server after the initial addr notification has been received by
the client and before the server sends the client the new primary
address. Then correct the way that the expected error condition is
detected by the test - when Dynamic Address Reconfiguration is foiled by
SELinux, then the client will never recieve the first "Client added"
notification, so exit with 52 in such case and check for this exit code
in the test.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>